WIP: session conversational view
- add sidebar
- layout changes in ui
