+++ agensight/ui/src/app/globals.css
@@ -130,57 +130,55 @@
     @apply text-3xl font-bold tracking-tight;
     line-height: 1.2;
   }
-  
+
   h2, .h2 {
     @apply text-2xl font-semibold tracking-tight;
     line-height: 1.3;
   }
-  
+
   h3, .h3 {
     @apply text-xl font-semibold tracking-tight;
     line-height: 1.4;
   }
-  
+
   h4, .h4 {
     @apply text-lg font-semibold tracking-tight;
     line-height: 1.5;
   }
-  
+
   p, .p {
     @apply text-base leading-7;
   }
-  
+
   .small {
     @apply text-sm leading-6;
   }
-  
+
   .tiny {
     @apply text-xs leading-5;
   }
-  
+
   .lead {
     @apply text-lg text-muted-foreground;
   }
-  
+
   .large {
     @apply text-lg font-medium;
   }
-  
+
   code {
     @apply bg-muted px-1.5 py-0.5 rounded text-sm font-mono;
   }
-  
+
   /* Definition Lists */
   dl dt {
     @apply text-sm font-medium leading-none;
   }
-  
+
   dl dd {
     @apply text-base text-muted-foreground;
   }
 }
-
-
 
 /* Custom utilities for the playground style */
 @layer utilities {
@@ -188,7 +186,7 @@
     filter: drop-shadow(0 4px 3px rgb(0 0 0 / 0.07)) drop-shadow(0 2px 2px rgb(0 0 0 / 0.06));
     transition: all 0.2s ease;
   }
-  
+
   .graph-node:hover {
     filter: drop-shadow(0 10px 8px rgb(0 0 0 / 0.04)) drop-shadow(0 4px 3px rgb(0 0 0 / 0.1));
     transform: translateY(-1px);
@@ -204,37 +202,37 @@
       stroke-dashoffset: 1000;
     }
   }
-  
+
   /* Custom scrollbar styles */
   .scroll-container::-webkit-scrollbar,
   pre::-webkit-scrollbar {
     width: 8px;
     height: 8px;
   }
-  
+
   .scroll-container::-webkit-scrollbar-track,
   pre::-webkit-scrollbar-track {
     background: transparent;
   }
-  
+
   .scroll-container::-webkit-scrollbar-thumb,
   pre::-webkit-scrollbar-thumb {
     background: var(--muted-foreground);
     opacity: 0.5;
     border-radius: 4px;
   }
-  
+
   .scroll-container::-webkit-scrollbar-thumb:hover,
   pre::-webkit-scrollbar-thumb:hover {
     background: var(--primary);
   }
-  
+
   /* Firefox scrollbar */
   .scroll-container, pre {
     scrollbar-width: thin;
     scrollbar-color: var(--muted-foreground) transparent;
   }
-  
+
   /* Glass morphism effect for panels */
   .glass-panel {
     background: rgba(255, 255, 255, 0.2);
@@ -242,18 +240,18 @@
     border: 1px solid rgba(255, 255, 255, 0.1);
     box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
   }
-  
+
   .dark .glass-panel {
     background: rgba(30, 41, 59, 0.7);
     border: 1px solid rgba(255, 255, 255, 0.05);
   }
-  
+
   /* Grid background */
   .grid-bg {
     background-size: 25px 25px;
     background-image: none;
   }
-  
+
   .dark .grid-bg {
     background-image: none;
   }
