+++ agensight/ui/src/components/HeaderWrapper.tsx
@@ -5,11 +5,9 @@ import { Header } from "@/components/Header";
 import { useTheme } from "@/components/ThemeProvider";
 
 export function HeaderWrapper({ children }: { children: React.ReactNode }) {
-  const { darkMode, toggleDarkMode } = useTheme();
   
   return (
     <div className="min-h-screen flex flex-col">
-      <Header darkMode={darkMode} toggleDarkMode={toggleDarkMode} />
       <main className="flex-1">
         {children}
       </main>
